cleaning up hypothesis engine - current code is very bloated
