+++ projects/orientation_tuning/config.yaml
@@ -24,6 +24,16 @@ inputs:
 
 
 # ========================
+# Experiment Runtime
+# ========================
+
+experiment_params:
+  x_min: 0.0
+  x_max: 6.283185307179586
+  n_bins: 100
+
+
+# ========================
 # Prompt Defaults
 # ========================
